+++ animation_nodes/ui/node_menu.py
@@ -445,6 +445,7 @@ class MeshOperatorsMenu(bpy.types.Menu):
         layout = self.layout
         insertNode(layout, "an_FindClosePointsNode", "Find Close Points")
         insertNode(layout, "an_EdgeToTubeNode", "Edge to Tube")
+        insertNode(layout, "an_MeshPointsScatterNode", "Scatter Points")
         layout.separator()
         insertNode(layout, "an_CreateEdgesNode", "Create Edges")
         insertNode(layout, "an_CreateEdgeIndicesNode", "Create Edge Indices")