+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added List to Element link conversion.
 - Added GP sockets link conversions.
 - Added input range to Remap Falloff node.
+- Added Mesh Points Scatter node.
 
 ### Fixed